+++ src/FrozenWasteland.cpp
@@ -22,6 +22,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelEverlastingGlottalStopper);
 	p->addModel(modelTheGardener);
 	p->addModel(modelHairPick);
+	//p->addModel(modelHoldingPattern);
 	p->addModel(modelJustAPhaser);
 	p->addModel(modelLissajousLFO);
 	p->addModel(modelMidichlorian);
@@ -51,7 +52,6 @@ void init(rack::Plugin *p) {
 	p->addModel(modelRouletteLFO);
 	p->addModel(modelSeriouslySlowEG);
 	p->addModel(modelSeriouslySlowLFO);
-	//p->addModel(modelTestDelay);
 	p->addModel(modelVoxInhumana);
 	p->addModel(modelVoxInhumanaExpander);
 	p->addModel(modelCDCSeriouslySlowLFO);